test(auto-reply): consolidate command suites
