+++ src/auto-reply/reply/commands-export-trajectory.test-support.ts
